kalshi_client.py
"""
Kalshi API client wrapper.
Handles authentication, requests, and retries.
"""
import time
import logging
from typing import List, Optional
import requests
from requests.adapters import HTTPAdapter
from urllib3.util.retry import Retry

from config import Config
from models import Market, Position

logger = logging.getLogger(__name__)


class KalshiClient:
    """Client for interacting with Kalshi API."""
    
    def __init__(self, config: Config):
        self.config = config
        self.base_url = config.kalshi_base_url
        self.api_key = config.kalshi_api_key
        self.api_secret = config.kalshi_api_secret
        self.session = self._create_session()
        self._token: Optional[str] = None
        self._token_expiry: float = 0
        
    def _create_session(self) -> requests.Session:
        """Create a requests session with retry strategy."""
        session = requests.Session()
        retry_strategy = Retry(
            total=3,
            backoff_factor=1,
            status_forcelist=[429, 500, 502, 503, 504],
            allowed_methods=["GET", "POST"]
        )
        adapter = HTTPAdapter(max_retries=retry_strategy)
        session.mount("http://", adapter)
        session.mount("https://", adapter)
        return session
    
    def _authenticate(self) -> str:
        """Authenticate with Kalshi API and return access token."""
        # TODO: Implement real Kalshi authentication
        # For now, return a mock token
        if self.config.mode == "SHADOW":
            logger.info("SHADOW mode: Using mock authentication")
            return "mock_token"
        
        # Real authentication would look like:
        # response = self.session.post(
        #     f"{self.base_url}/login",
        #     json={
        #         "email": self.api_key,  # or username
        #         "password": self.api_secret
        #     }
        # )
        # response.raise_for_status()
        # data = response.json()
        # return data["token"]
        
        raise NotImplementedError("Real Kalshi authentication not yet implemented")
    
    def _get_token(self) -> str:
        """Get valid access token, refreshing if needed."""
        if not self._token or time.time() >= self._token_expiry:
            self._token = self._authenticate()
            self._token_expiry = time.time() + 3600  # Assume 1 hour expiry
        return self._token
    
    def _request(self, method: str, endpoint: str, **kwargs) -> dict:
        """Make authenticated request to Kalshi API."""
        if self.config.mode == "SHADOW":
            logger.debug(f"SHADOW mode: Would {method} {endpoint}")
            return {}
        
        token = self._get_token()
        headers = kwargs.pop("headers", {})
        headers["Authorization"] = f"Bearer {token}"
        headers["Content-Type"] = "application/json"
        
        url = f"{self.base_url}/{endpoint.lstrip('/')}"
        response = self.session.request(method, url, headers=headers, **kwargs)
        response.raise_for_status()
        return response.json()
    
    def get_account_balance(self) -> float:
        """
        Get current account balance (bankroll).
        
        Returns:
            Current cash balance in dollars.
        """
        # TODO: Implement real Kalshi balance endpoint
        # For now, return a mock balance
        if self.config.mode == "SHADOW":
            logger.debug("SHADOW mode: Returning mock balance")
            return 10000.0  # Mock $10k bankroll
        
        # Real implementation:
        # data = self._request("GET", "/portfolio/balance")
        # return float(data["balance"])
        
        raise NotImplementedError("Real Kalshi balance endpoint not yet implemented")
    
    def get_positions(self) -> List[Position]:
        """
        Get all open positions.
        
        Returns:
            List of Position objects.
        """
        # TODO: Implement real Kalshi positions endpoint
        if self.config.mode == "SHADOW":
            logger.debug("SHADOW mode: Returning empty positions")
            return []
        
        # Real implementation:
        # data = self._request("GET", "/portfolio/positions")
        # positions = []
        # for pos_data in data["positions"]:
        #     positions.append(Position(
        #         market_id=pos_data["market_id"],
        #         game_id=pos_data.get("game_id", ""),
        #         team=pos_data.get("team", ""),
        #         league=pos_data.get("league", ""),
        #         quantity=int(pos_data["quantity"]),
        #         average_price=float(pos_data["average_price"]),
        #         current_yes_price=float(pos_data.get("current_yes_price", 0)),
        #         unrealized_pnl=float(pos_data.get("unrealized_pnl", 0)),
        #         max_loss=float(pos_data.get("max_loss", 0))
        #     ))
        # return positions
        
        raise NotImplementedError("Real Kalshi positions endpoint not yet implemented")
    
    def fetch_sports_markets(self) -> List[Market]:
        """
        Fetch all open sports markets.
        
        Returns:
            List of Market objects.
        """
        # TODO: Implement real Kalshi markets endpoint
        # For now, return mock markets for testing
        if self.config.mode == "SHADOW":
            logger.debug("SHADOW mode: Returning mock markets")
            from datetime import datetime, timedelta
            import random
            
            # Generate some mock markets
            mock_markets = []
            teams = [
                ("Lakers", "Warriors", "NBA"),
                ("Chiefs", "Bills", "NFL"),
                ("Man City", "Liverpool", "EPL"),
            ]
            
            for i, (team_a, team_b, league) in enumerate(teams):
                game_id = f"{league}_{team_a}_{team_b}_{i}"
                # Mock prices with some edge opportunities
                yes_price = 0.45 + random.uniform(-0.1, 0.1)
                no_price = 1.0 - yes_price
                spread = random.uniform(0.02, 0.10)
                
                mock_markets.append(Market(
                    market_id=f"market_{i}",
                    event_name=f"{team_a} vs {team_b}",
                    game_id=game_id,
                    league=league,
                    team=team_a,
                    best_yes_price=max(0.01, min(0.99, yes_price)),
                    best_no_price=max(0.01, min(0.99, no_price)),
                    volume=random.randint(1000, 10000),
                    spread=spread,
                    start_time=datetime.now() + timedelta(hours=2),
                    settlement_time=datetime.now() + timedelta(hours=4),
                    title=f"{team_a} to win vs {team_b}"
                ))
            
            return mock_markets
        
        # Real implementation:
        # data = self._request("GET", "/markets", params={"category": "sports"})
        # markets = []
        # for market_data in data["markets"]:
        #     # Parse market data and create Market objects
        #     # This will need to be adjusted based on actual Kalshi API response format
        #     markets.append(Market(...))
        # return markets
        
        raise NotImplementedError("Real Kalshi markets endpoint not yet implemented")
    
    def place_yes_order(
        self, 
        market_id: str, 
        price: float, 
        size: int
    ) -> Optional[str]:
        """
        Place a limit order for YES contracts.
        
        Args:
            market_id: Market identifier
            price: Limit price (0-1)
            size: Number of contracts
            
        Returns:
            Order ID if successful, None otherwise
        """
        if self.config.mode == "SHADOW":
            logger.info(
                f"SHADOW mode: Would place YES order - "
                f"market={market_id}, price={price:.4f}, size={size}"
            )
            return f"shadow_order_{market_id}_{int(time.time())}"
        
        # Real implementation:
        # data = self._request(
        #     "POST",
        #     "/orders",
        #     json={
        #         "market_id": market_id,
        #         "side": "yes",
        #         "action": "buy",
        #         "type": "limit",
        #         "price": price,
        #         "count": size
        #     }
        # )
        # return data.get("order_id")
        
        raise NotImplementedError("Real Kalshi order placement not yet implemented")

